Refactored Dialog in Dialog and ExtendedDialog, FormCheckbox
