+++ resources/js/components/twc-ui/form-checkbox.tsx
@@ -1,0 +1,4 @@
+import { Checkbox, type CheckboxProps } from './checkbox'
+
+export { Checkbox as FormCheckbox }
+export type { CheckboxProps as FormCheckboxProps }
