+++ resources/js/components/twc-ui/popover.tsx
@@ -3,10 +3,10 @@
 import * as React from 'react'
 import {
   Dialog as AriaDialog,
-  DialogProps as AriaDialogProps,
+  type DialogProps as AriaDialogProps,
   DialogTrigger as AriaDialogTrigger,
   Popover as AriaPopover,
-  PopoverProps as AriaPopoverProps,
+  type PopoverProps as AriaPopoverProps,
   composeRenderProps
 } from 'react-aria-components'
 
@@ -21,11 +21,11 @@ const Popover = ({ className, offset = 4, ...props }: AriaPopoverProps) => (
       cn(
         'z-50 rounded-md border bg-popover text-popover-foreground shadow-md outline-none',
         /* Entering */
-        'data-[entering]:animate-in data-[entering]:fade-in-0 data-[entering]:zoom-in-95',
+        'entering:fade-in-0 entering:zoom-in-95 entering:animate-in',
         /* Exiting */
-        'data-[exiting]:animate-out data-[exiting]:fade-out-0 data-[exiting]:zoom-out-95',
+        'exiting:fade-out-0 exiting:zoom-out-95 exiting:animate-out',
         /* Placement */
-        'data-[placement=bottom]:slide-in-from-top-2 data-[placement=left]:slide-in-from-right-2 data-[placement=right]:slide-in-from-left-2 data-[placement=top]:slide-in-from-bottom-2',
+        'placement-bottom:slide-in-from-top-2 placement-left:slide-in-from-right-2 placement-right:slide-in-from-left-2 placement-top:slide-in-from-bottom-2',
         className
       )
     )}
