+++ resources/js/components/twc-ui/registry/dialog.json
@@ -5,7 +5,7 @@
   "title": "Dialog",
   "description": "Dialog component",
   "dependencies": ["class-variance-authority", "react-aria-components", "lucide-react"],
-  "registryDependencies": ["@twc-ui/alert-dialog", "@twc-ui/button"],
+  "registryDependencies": ["@twc-ui/button"],
   "files": [
     {
       "path": "resources/js/components/twc-ui/dialog.tsx",
