+++ resources/js/components/twc-ui/registry/form-checkbox.json
@@ -1,0 +1,15 @@
+{
+  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
+  "name": "form-checkbox",
+  "type": "registry:ui",
+  "title": "FormCheckbox",
+  "description": "FormCheckbox component with form integration",
+  "registryDependencies": ["@twc-ui/checkbox"],
+  "files": [
+    {
+      "path": "resources/js/components/twc-ui/form-checkbox.tsx",
+      "type": "registry:ui",
+      "target": "components/twc-ui/form-checkbox.tsx"
+    }
+  ]
+}
