+++ resources/js/docs/components/react-aria/checkbox.mdx
@@ -3,8 +3,3 @@ title: Checkbox
 ---
 <Installation registry="checkbox" />
 
-## Components
-### Checkbox
-#### Props
-### CheckboxGroup
-Not implemented yet.
