+++ resources/js/docs/components/twiceware-ui-forms/form-checkbox.mdx
@@ -1,0 +1,5 @@
+---
+title: FormCheckbox
+---
+# Installation
+<Installation registry="form-checkbox" />
