+++ resources/js/docs/components/twiceware-ui/alert-dialog.mdx
@@ -27,5 +27,5 @@ const doSomething = async () => {
 | title             | `string`                                                                | ``            | Title of the alert dialog                       |
 | message           | `string`                                                                | ``            | Message                                         |
 | buttonTitle       | `string`                                                                | `OK`          | Label of the OK button                          |
-| cancelButtonTitle | `string`                                                                | `Cancel`      | Label of the cancel button                      |
+| cancelButtonTitle | `string`                                                                | EN: `Cancel`<br/>DE: `Abbruch`      | Label of the cancel button                      |
 | variant           | <TypeInfo title="variant" types={['default', 'destructive', 'info']} /> | `destructive` | Icon of the dialog and variant of the OK-button |
