--- resources/js/docs/components/twiceware-ui/form-checkbox.mdx
@@ -1,9 +1,0 @@
----
-title: FormCheckbox
----
-
-## Components
-### Checkbox
-#### Props
-### CheckboxGroup
-Not implemented yet.
