+++ resources/js/hooks/use-docs-structure.tsx
@@ -49,7 +49,8 @@ const sortDocsItems = (items: DocItem[]): DocItem[] => {
     'quick-start': 4,
     components: 10,
     'twiceware-ui': 1,
-    'react-aria': 2,
+    'twiceware-ui-forms': 2,
+    'react-aria': 3,
     hooks: 15,
     api: 20,
     examples: 30,
